feat: 오늘의 카드 추천 기능 구현

+++ src/main/java/site/benepay/domain/benefit/dto/NearbyBenefitResponseDto.java
@@ -24,5 +24,5 @@ public class NearbyBenefitResponseDto {
 	 */
 	private String benefitSummary;
 
-	private Double distanceMeters;
+	private Long distanceMeters;
 }
